Simplified .npy key and preprocessing for inference
PreProcessing/PreProcessingTools.py -> Removed getTilesFromNonBackground that was used for inference preprocessing, as background is now a class. It is replaced by getAllTiles method to create the .npy with full tessellation.

Inference/Preprocess.py ->. updated to reflect the above changes.

QA/Normalization/Colour/train_macenko -> adjusted the example slide.

Utils/npyExportTools -> changed the name of the key used to score config and dataframes to npy - simplified to only Patch Size and Visibility levels.